+++ drivers/media/platform/rockchip-vpu/rockchip_vpu_hw.c
@@ -410,8 +410,8 @@ void rockchip_vpu_vp8e_assemble_bitstream(struct rockchip_vpu_ctx *ctx,
 	if (WARN_ON(dst_buf->vp8e.dct_offset + dct_size > dst_size))
 		return;
 
-	vpu_debug(1, "%s: hdr_size = %d, ext_hdr_size = %d, dct_size = %d\n",
-			__func__, hdr_size, ext_hdr_size, dct_size);
+	vpu_debug(1, "%s: hdr_size = %ld, ext_hdr_size = %ld, dct_size = %ld\n",
+			__func__, (long) hdr_size, (long) ext_hdr_size, (long) dct_size);
 
 	memmove(dst + hdr_size + ext_hdr_size,
 		dst + dst_buf->vp8e.dct_offset, dct_size);
@@ -452,8 +452,8 @@ void rockchip_vpu_h264e_assemble_bitstream(struct rockchip_vpu_ctx *ctx,
 	if (WARN_ON(sps_size + pps_size + slices_size > dst_size))
 		return;
 
-	vpu_debug(1, "%s: sps_size = %u, pps_size = %u, slices_size = %u\n",
-		__func__, sps_size, pps_size, slices_size);
+	vpu_debug(1, "%s: sps_size = %lu, pps_size = %lu, slices_size = %lu\n",
+		__func__, (long) sps_size, (long) pps_size, (long) slices_size);
 
 	memcpy(dst, sps->buffer, sps_size);
 	memcpy(dst + sps_size, pps->buffer, pps_size);
